ch(fastfoodfast): optimise code
remove unused lines of code

[Delivers #160982299]
