+++ .travis.yml
@@ -16,19 +16,15 @@ addons:
     repo_token:
       secure: "d+dccm3DDJnuOfdvFCswDZUs0KLNTY0LutweQWX4XxAgjgl/br3LqnrlnrsuBrw4AKiQSxXmnr/CsGUx1vgVDpO0j5i2pDIEwHFEjrby8aNCm5B1mv+q7El+RQu5hnjCmtReheJO/bIKRm8Mb7NqZJPZ5sbsPT1KBMatkU0QAAe7NEPgOUcmVvr7/iL4r9XZtoclJODAHQTBz1l4XXTiSg3JXahHE+TbQ8znjV+Tl2L1xybgyZePDCrPrDhlGqjJMaqojKMg1xLriVldBZlUQamvYSBuSNclWwxHNG7j7VQHkyGwfsef/4ti4KlibnBFmFC3TbWvIzFF/hsMCGv7PHDcyrVsPgYH6MGrsBCygtWa9VGHL1De4H87XCCBuIw6NYuxEgEOPaEiEhD9FYJiGyqowXq05yH2JYsvm2mpvNkLnenQ3YAyfuCg+kJaC0t8osyZaaFjp8vJYVO/AHN3SC6h7vUS3G6XqWipXDIx1xFuP+zLJGOs3/KLwI+/8Ae3sej1168pvKw/xmEx14g+/iCNI3PlP9jI66a6M4xb7jFjOx5685BrXDgIFWpbF7C7eB4PFun3pzChl71+EyXB2j5IRXunTuHgFqK5TWkYAORkwYz/OsZn+rZclY1RuJatvGVaeyq/t5MEauB4+Y4zy5f6MWJNflcZiM0rkITibJY="
 
-before_script:  # code coverage tool
+before_script:  
   - curl -L https://codeclimate.com/downloads/test-reporter/test-reporter-latest-linux-amd64 > ./cc-test-reporter
   - chmod +x ./cc-test-reporter
   - ./cc-test-reporter before-build
 
 script:
-#  - pylint --output-format=text *.py
   - pytest
   - coverage run --source api --branch -m unittest discover
-  # - coverage erase
-  # - coverage run test_api.py && coverage html
   - coverage report -m
-  # submit the reports
   - coveralls
 
 after_script:
