+++ api/docs/sample.yml
@@ -1,0 +1,25 @@
+Create item
+---
+tags:
+- MethodView
+parameters:
+- name: Item
+  in: body
+  schema:
+    type: object
+    required:
+    - name
+    - type
+    properties:
+      name:
+        type: string
+        maxLength: 31
+        description: global unique
+      type:
+        type: string
+        maxLength: 10
+responses:
+  201:
+    description: OK
+    schema:
+      $ref: '#/definitions/Item'
