+++ api/views/userview.py
@@ -1,26 +1,10 @@
-from api.user import User
-from api.order import Order
+from api.models.user import User
+from api.models.order import Order
 from api import app
 from api.views.decorators import *
 
 USER = User()
 ORDER = Order()
-
-@app.route('/api/v1/users', methods=['GET'])
-def get_all_users():
-    """ A route to return all of the available users. """
-    return jsonify({'users': USER.fetch_all_users()})
-
-@app.route('/api/v1/users/<int:user_id>', methods=['GET'])
-def get_user(user_id):
-    """ Get a specific user with given id."""
-    user = USER.get_user(user_id)
-    return jsonify({'user': user})
-
-@app.route('/api/v1/users/orders/<int:order_id>', methods=['PUT'])
-def update_user_order(order_id):
-    """ update order details with put request. """
-    return jsonify({'order': ORDER.update_user_order(order_id, request.json)})
 
 @app.route('/api/v1/users/orders', methods=['GET'])
 @jwt_required
@@ -51,14 +35,56 @@ def get_user_orders():
     else:
         return jsonify({"message":"no orders available for current user"})
 
-@app.route('/api/v1/users/<int:user_id>', methods=['PUT'])
-def elevate_user_to_admin(user_id):
-    """ update user to admin. """
-    status = USER.assign_admin_privileges(user_id)
-    return jsonify(status)
 
-@app.route('/api/v1/users/admin/<string:email>', methods=['GET'])
-def get_user_data(email):
-    """ Get orders for a specific user."""
-    data = USER.get_user_data_from(email)
-    return jsonify(data)
+@app.route('/api/v1/users/orders', methods=['POST'])
+@jwt_required
+def create_order():
+    """
+        Create a new order
+        Allows a customer post an order
+        ---
+        tags:
+          - ORDER
+        parameters:
+          - in: body
+            item: body
+            quantity: body
+            schema:
+              id: Order
+              required:
+                - item
+                - quantity
+              properties:
+                item:
+                  type: string
+                  description: ordered item
+                  default: Chips
+                quantity:
+                  type: integer
+                  description: number of items requested
+                  default: 1
+        responses:
+          201:
+            description: New order created
+        """
+    if not request.json or not 'item' in request.json:
+        return jsonify({'error': 'Missing Item parameter in request'}), 400
+    try:
+        request.json['quantity'] = int(request.json['quantity'])
+    except ValueError:
+        return jsonify({'error': 'Bad request'}), 400
+
+    current_user = get_jwt_identity()
+    logged_in_user = current_user['id']
+
+    order = ORDER.check_if_order_exists(logged_in_user , \
+                                        request.json['item'], request.json['quantity'])
+    if order:
+        return jsonify({'error': 'Order already exists'}), 409
+    else:
+        create_user_order = ORDER.create_order(logged_in_user, request.json)
+        if create_user_order:
+            return jsonify({"message":"Order successfully created",
+                            'id': create_user_order}), 201
+        else:
+            return jsonify({'error': True, "message":"Unable to support request"}), 400
